Editor Sidebar: Rename `util.js` -> `constants.js`
